+++ src/hoard.rs
@@ -128,7 +128,16 @@ impl Hoard {
             ("new", Some(_sub_m)) => {
                 let default_namespace = self.config.as_ref().unwrap().default_namespace.clone();
                 let new_command = HoardCommand::default()
-                    .with_command_string_input(None)
+                    .with_command_string_input(
+                        None,
+                        &self
+                            .config
+                            .as_ref()
+                            .unwrap()
+                            .parameter_token
+                            .clone()
+                            .unwrap(),
+                    )
                     .with_namespace_input(Some(default_namespace))
                     .with_name_input(None, &self.trove)
                     .with_description_input(None)
@@ -147,14 +156,16 @@ impl Hoard {
                 } else {
                     match commands_gui::run(&mut self.trove, self.config.as_ref().unwrap()) {
                         Ok(selected_command) => {
-                            // Is set if a command is selected in GUI
-                            if !selected_command.is_empty() {
-                                //TODO: If run as cli program, copy command into clipboard, else will be written to READLINE_LINE
-                                autocomplete_command = selected_command;
+                            if let Some(c) = selected_command {
+                                // Is set if a command is selected in GUI
+                                if !c.command.is_empty() {
+                                    //TODO: If run as cli program, copy command into clipboard, else will be written to READLINE_LINE
+                                    autocomplete_command = c.command;
+                                }
                             }
                         }
-                        Err(error) => {
-                            println!("{}", error);
+                        Err(e) => {
+                            println!("{}", e);
                         }
                     }
                 }
@@ -162,7 +173,9 @@ impl Hoard {
             // Load command by name into clipboard, if available
             ("pick", Some(sub_m)) => {
                 if let Some(command_name) = sub_m.value_of("name") {
-                    let command_result = self.trove.pick_command(command_name);
+                    let command_result = self
+                        .trove
+                        .pick_command(self.config.as_ref().unwrap(), command_name);
                     match command_result {
                         Ok(c) => {
                             println!("{}", c.command);
@@ -239,12 +252,23 @@ impl Hoard {
             ("edit", Some(sub_m)) => {
                 if let Some(command_name) = sub_m.value_of("name") {
                     println!("Editing {:?}", command_name);
-                    let command_to_edit = self.trove.pick_command(command_name);
+                    let command_to_edit = self
+                        .trove
+                        .pick_command(self.config.as_ref().unwrap(), command_name);
                     match command_to_edit {
                         Ok(c) => {
                             println!("{}", c.command);
                             let new_command = HoardCommand::default()
-                                .with_command_string_input(Some(c.command.clone()))
+                                .with_command_string_input(
+                                    Some(c.command.clone()),
+                                    &self
+                                        .config
+                                        .as_ref()
+                                        .unwrap()
+                                        .parameter_token
+                                        .clone()
+                                        .unwrap(),
+                                )
                                 .with_name_input(Some(c.name.clone()), &self.trove)
                                 .with_description_input(c.description.clone())
                                 .with_tags_input(Some(c.tags_as_string()))